requirements.txt
deepagents
langchain-core
langchain-google-genai
google-auth
python-dotenv
langgraph-checkpoint-postgres
psycopg[binary]
langchain-community
duckduckgo-search
tavily-python
fastapi
uvicorn[standard]
